# Simplify lex_identifier