--- a/src/lexer.jl
+++ b/src/lexer.jl
@@ -408,7 +408,6 @@ function next_token(l::Lexer, start = true)
         readon(l)
         return lex_cmd(l);
     elseif is_identifier_start_char(c)
-        readon(l)
         return lex_identifier(l, c)
     elseif isdigit(c)
         readon(l)
@@ -1024,305 +1023,83 @@ function lex_cmd(l::Lexer, doemit=true)
     end
 end
 
-function tryread(l, str, k, c)
-    for s in str
-        c = peekchar(l)
-        if c != s
-            if !is_identifier_char(c)
-                return emit(l, IDENTIFIER)
-            end
-            return readrest(l, c)
-        else
-            readchar(l)
-        end
+function lex_identifier(l::Lexer{IO_t,T}, c) where {IO_t,T}
+    if T == Token
+        readon(l)
     end
-    if is_identifier_char(peekchar(l))
-        return readrest(l, c)
-    end
-    return emit(l, k)
-end
-
-function readrest(l, c)
+    cnt = 1
+    h = simple_hash(Int(c), cnt, 0)
     while true
         pc, ppc = dpeekchar(l)
         if !is_identifier_char(pc) || (pc == '!' && ppc == '=')
             break
         end
         c = readchar(l)
+        cnt += 1
+        h = simple_hash(Int(c), cnt, h)
     end
 
-    return emit(l, IDENTIFIER)
+    return emit(l, get(kw_hash, h, IDENTIFIER))
 end
 
+function simple_hash(c, cnt, h)
+    h = h*c + c + cnt
+end
 
-function _doret(l, c)
-    if !is_identifier_char(c)
-        return emit(l, IDENTIFIER)
-    else
-        return readrest(l, c)
+function simple_hash(str)
+    ind = 1
+    cnt = 1
+    h = 0
+    while ind <= length(str)
+        h = simple_hash(Int(str[ind]), cnt, h)
+        cnt += 1
+        ind = nextind(str, ind)
     end
+    h
 end
 
-function lex_identifier(l, c)
-    if c == 'a'
-        return tryread(l, ('b', 's', 't', 'r', 'a', 'c', 't'), ABSTRACT, c)
-    elseif c == 'b'
-        c = peekchar(l)
-        if c == 'a'
-            c = readchar(l)
-            return tryread(l, ('r', 'e', 'm', 'o', 'd', 'u', 'l', 'e'), BAREMODULE, c)
-        elseif c == 'e'
-            c = readchar(l)
-            return tryread(l, ('g', 'i', 'n'), BEGIN, c)
-        elseif c == 'r'
-            c = readchar(l)
-            return tryread(l, ('e', 'a', 'k'), BREAK, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'c'
-        c = peekchar(l)
-        if c == 'a'
-            c = readchar(l)
-            return tryread(l, ('t', 'c', 'h'), CATCH, c)
-        elseif c == 'o'
-            readchar(l)
-            c = peekchar(l)
-            if c == 'n'
-                readchar(l)
-                c = peekchar(l)
-                if c == 's'
-                    readchar(l)
-                    c = peekchar(l)
-                    return tryread(l, ('t',), CONST, c)
-                elseif c == 't'
-                    readchar(l)
-                    c = peekchar(l)
-                    return tryread(l, ('i', 'n', 'u', 'e'), CONTINUE, c)
-                else
-                    return _doret(l, c)
-                end
-            else
-                return _doret(l, c)
-            end
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'd'
-        return tryread(l, ('o'), DO, c)
-    elseif c == 'e'
-        c = peekchar(l)
-        if c == 'l'
-            readchar(l)
-            c = peekchar(l)
-            if c == 's'
-                readchar(l)
-                c = peekchar(l)
-                if c == 'e'
-                    readchar(l)
-                    c = peekchar(l)
-                    if !is_identifier_char(c)
-                        return emit(l, ELSE)
-                    elseif c == 'i'
-                        c = readchar(l)
-                        return tryread(l, ('f'), ELSEIF ,c)
-                    else
-                        return _doret(l, c)
-                    end
-                else
-                    return _doret(l, c)
-                end
-            else
-                return _doret(l, c)
-            end
-        elseif c == 'n'
-            c = readchar(l)
-            return tryread(l, ('d'), END, c)
-        elseif c == 'x'
-            c = readchar(l)
-            return tryread(l, ('p', 'o', 'r', 't'), EXPORT, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'f'
-        c = peekchar(l)
-        if c == 'a'
-            c = readchar(l)
-            return tryread(l, ('l', 's', 'e'), FALSE, c)
-        elseif c == 'i'
-            c = readchar(l)
-            return tryread(l, ('n', 'a', 'l', 'l', 'y'), FINALLY, c)
-        elseif c == 'o'
-            c = readchar(l)
-            return tryread(l, ('r'), FOR, c)
-        elseif c == 'u'
-            c = readchar(l)
-            return tryread(l, ('n', 'c', 't', 'i', 'o', 'n'), FUNCTION, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'g'
-        return tryread(l, ('l', 'o', 'b', 'a', 'l'), GLOBAL, c)
-    elseif c == 'i'
-        c = peekchar(l)
-        if c == 'f'
-            readchar(l)
-            c = peekchar(l)
-            if !is_identifier_char(c)
-                return emit(l, IF)
-            else
-                return readrest(l, c)
-            end
-        elseif c == 'm'
-            readchar(l)
-            c = peekchar(l)
-            if c == 'p'
-                readchar(l)
-                c = peekchar(l)
-                if c == 'o'
-                    readchar(l)
-                    c = peekchar(l)
-                    if c == 'r'
-                        readchar(l)
-                        c = peekchar(l)
-                        if c == 't'
-                            readchar(l)
-                            c = peekchar(l)
-                            if !is_identifier_char(c)
-                                return emit(l, IMPORT)
-                            elseif c == 'a'
-                                c = readchar(l)
-                                return tryread(l, ('l','l'), IMPORTALL, c)
-                            else
-                                return _doret(l, c)
-                            end
-                        else
-                            return _doret(l, c)
-                        end
-                    else
-                        return _doret(l, c)
-                    end
-                else
-                    return _doret(l, c)
-                end
-            else
-                return _doret(l, c)
-            end
-        elseif c == 'n'
-            readchar(l)
-            c = peekchar(l)
-            if !is_identifier_char(c)
-                return emit(l, IN)
-            else
-                return readrest(l, c)
-            end
-        elseif (@static VERSION >= v"0.6.0-dev.1471" ? true : false) && c == 's'
-            c = readchar(l)
-            return tryread(l, ('a'), ISA, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'l'
-        c = peekchar(l)
-        if c == 'e'
-            readchar(l)
-            return tryread(l, ('t'), LET, c)
-        elseif c == 'o'
-            readchar(l)
-            return tryread(l, ('c', 'a', 'l'), LOCAL, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'm'
-        c = peekchar(l)
-        if c == 'a'
-            c = readchar(l)
-            return tryread(l, ('c', 'r', 'o'), MACRO, c)
-        elseif c == 'o'
-            c = readchar(l)
-            return tryread(l, ('d', 'u', 'l', 'e'), MODULE, c)
-        elseif c == 'u'
-            c = readchar(l)
-            return tryread(l, ('t', 'a', 'b', 'l', 'e'), MUTABLE, c)
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'o'
-        return tryread(l, ('u', 't', 'e', 'r'), OUTER, c)
-    elseif c == 'p'
-        return tryread(l, ('r', 'i', 'm', 'i', 't', 'i', 'v', 'e'), PRIMITIVE, c)
-    elseif c == 'q'
-        return tryread(l, ('u', 'o', 't', 'e'), QUOTE, c)
-    elseif c == 'r'
-        return tryread(l, ('e', 't', 'u', 'r', 'n'), RETURN, c)
-    elseif c == 's'
-        return tryread(l, ('t', 'r', 'u', 'c', 't'), STRUCT, c)
-    elseif c == 't'
-        c = peekchar(l)
-        if c == 'r'
-            readchar(l)
-            c = peekchar(l)
-            if c == 'u'
-                c = readchar(l)
-                return tryread(l, ('e'), TRUE, c)
-            elseif c == 'y'
-                readchar(l)
-                c = peekchar(l)
-                if !is_identifier_char(c)
-                    return emit(l, TRY)
-                else
-                    c = readchar(l)
-                    return _doret(l, c)
-                end
-            else
-                return _doret(l, c)
-            end
-        elseif c == 'y'
-            readchar(l)
-            c = peekchar(l)
-            if c == 'p'
-                readchar(l)
-                c = peekchar(l)
-                if c == 'e'
-                    readchar(l)
-                    c = peekchar(l)
-                    if !is_identifier_char(c)
-                        return emit(l, TYPE)
-                    else
-                        c = readchar(l)
-                        return _doret(l, c)
-                    end
-                else
-                    return _doret(l, c)
-                end
-            else
-                return _doret(l, c)
-            end
-        else
-            return _doret(l, c)
-        end
-    elseif c == 'u'
-        return tryread(l, ('s', 'i', 'n', 'g'), USING, c)
-    elseif c == 'w'
-        c = peekchar(l)
-        if c == 'h'
-            readchar(l)
-            c = peekchar(l)
-            if c == 'e'
-                c = readchar(l)
-                return tryread(l, ('r', 'e'), WHERE, c)
-            elseif c == 'i'
-                c = readchar(l)
-                return tryread(l, ('l', 'e'), WHILE, c)
-            else
-                return _doret(l, c)
-            end
-        else
-            return _doret(l, c)
-        end
-    else
-        return _doret(l, c)
-    end
-end
+kws = [
+Tokens.ABSTRACT,
+Tokens.BAREMODULE,
+Tokens.BEGIN,
+Tokens.BREAK,
+Tokens.CATCH,
+Tokens.CONST,
+Tokens.CONTINUE,
+Tokens.DO,
+Tokens.ELSE,
+Tokens.ELSEIF,
+Tokens.END,
+Tokens.EXPORT,
+Tokens.FINALLY,
+Tokens.FOR,
+Tokens.FUNCTION,
+Tokens.GLOBAL,
+Tokens.IF,
+Tokens.IMPORT,
+Tokens.IMPORTALL,
+Tokens.LET,
+Tokens.LOCAL,
+Tokens.MACRO,
+Tokens.MODULE,
+Tokens.MUTABLE,
+Tokens.NEW,
+Tokens.OUTER,
+Tokens.PRIMITIVE,
+Tokens.QUOTE,
+Tokens.RETURN,
+Tokens.STRUCT,
+Tokens.TRY,
+Tokens.TYPE,
+Tokens.USING,
+Tokens.WHILE,
+Tokens.IN,
+Tokens.ISA,
+Tokens.WHERE,
+Tokens.TRUE,
+Tokens.FALSE,
+]
+
+const kw_hash = Dict(simple_hash(lowercase(string(kw))) => kw for kw in kws)
 
 end # module