Alternative to https://github.com/JuliaLang/Tokenize.jl/pull/187. Simplifies (and speeds up?) lex_identifier by building a hash as each character of an identifier is read and then checking against a table of keywords/word-operators before emiting. However, I'm not sure about the validity of the `simple_hash` function I've used.

Seems to cut timings on the first run to a little more than half with a neglible impact on once-compiled run time for the test below
```julia
for i = 1:5
    run(`julia -e "using Tokenize;@time @eval collect(tokenize(\"foo + bar\"));@time @eval collect(tokenize(\"foo + bar\"));"`)
end
```