@KristofferC 2021-08-10
Nice, https://github.com/JuliaLang/Tokenize.jl/pull/187 still has a quite significant effect though (0.67s -> 0.39s).