@pfitzseb 2021-08-10
So why not both? :)