@KristofferC 2021-08-10
Seems like the last commit made it worse?