@KristofferC 2021-08-11
If you don't use eval you don't get the full compilation time (things might start compiling before the timing starts).