@pfitzseb 2021-08-24
A few timings on my machine:

master:
```
  1.403908 seconds (157.57 k allocations: 8.206 MiB, 100.15% compilation time)
  0.000128 seconds (66 allocations: 4.891 KiB)
```
This PR:
```
  0.732055 seconds (149.31 k allocations: 8.002 MiB, 100.30% compilation time)
  0.000111 seconds (68 allocations: 4.922 KiB)
```
This PR + #187:
```
  0.421957 seconds (149.31 k allocations: 8.001 MiB, 100.53% compilation time)
  0.000132 seconds (68 allocations: 4.922 KiB)
```

So let's get this in ASAP, I think! :)